CORS configured for all controller classes. global api endpoints declared
